fixed_types_secure.hpp
This is secure version of fixed_types.hpp. It has static_assertions added and size_t variable explicitly used to make sure that there are no unintended consequences in the resulting program. If one of the fixed-width aliases ever happen to be different size than intended, static_assertion(s) make sure that the code does not compile. decltype(sizeof(unsigned)) is replaced with size_t to explicitly state that return type of sizeof operator is alias of size_t.  Extra inclusion of header files and usage of static_assertions increases size of this secure version a bit more.